+++ CHANGES.md
@@ -1,8 +1,7 @@
 #### [unreleased]
 * add most of `move_dir()` and `is_virtualbox()`
-* use `move_dir_fallback()` which is recursive copy > unlink, core `copy_dir()` doesn't work with how Git Updater renames downloads, can result in endless loop
 * update above to work with Rollback
-* put `move_dir()`, `move_dir_fallback()`, `is_virtualbox()` in `class Shim` for simpler usage from core functions
+* put `move_dir()` and `is_virtualbox()` in `class Shim` for simpler usage from core functions
 
 #### 10.9.0 / 2022-04-04
 * revert usage of `move_dir()` and `is_virtualbox()` -- for now